remove smart ami logic, because what if its deleted?
I don't want this logic in here. Otherwise when a Windows AMI that we
use is marked private by the owner (in this case Amazon), the plan
errors and does not let me proceed. If I do specify an existing AMI ID,
it forces the instance to be replaced...NOT ideal.

While this does remove some of the helpful logic, I see it doing more
harm than good.